+++ jvm/src/main/scala/gpdviz/server/GpdvizServiceImpl.scala
@@ -140,6 +140,8 @@ trait GpdvizServiceImpl extends GpdvizJsonImplicits with Logging {
       template
         .replace("#sysid", sysid)
         .replace("#externalUrl", cfg.externalUrl)
+        .replace("#pusher", cfg.pusher.map(_ ⇒
+          """<script src="//js.pusher.com/3.2/pusher.min.js"></script>""").getOrElse(""))
     }
 
     logger.debug(s"getSensorSystemIndex calling getSensorSystem sysid=$sysid")